fix: 가격 정보 없는 날짜 클릭 시 예외처리
